// src/db/generated.ts
/**
 * This file was generated by kysely-codegen.
 * Please do not edit it manually.
 */

import type { ColumnType } from "kysely";

export type CampaignStatus = "available" | "not_available";

export type CouponStatus = "available" | "not_available";

export type Generated<T> = T extends ColumnType<infer S, infer I, infer U>
  ? ColumnType<S, I | undefined, U>
  : ColumnType<T, T | undefined, T>;

export type Timestamp = ColumnType<Date, Date | string, Date | string>;

export type UserRole = "admin" | "user";

export interface Campaigns {
  created_at: Generated<Timestamp>;
  description: string | null;
  end_timestamp: Timestamp | null;
  id: Generated<string>;
  max_redemptions: number | null;
  name: string;
  redemptions_count: Generated<number>;
  start_timestamp: Timestamp;
  status: CampaignStatus;
  updated_at: Generated<Timestamp>;
}

export interface Coupons {
  campaign_id: string;
  code: string;
  created_at: Generated<Timestamp>;
  expiration_timestamp: Timestamp | null;
  id: Generated<string>;
  max_redemptions: number | null;
  redemptions_count: Generated<number>;
  status: CouponStatus;
  updated_at: Generated<Timestamp>;
}

export interface Redemptions {
  coupon_id: string;
  id: Generated<string>;
  redeemed_at: Generated<Timestamp>;
  updated_at: Generated<Timestamp>;
  user_id: string;
}

export interface Users {
  created_at: Generated<Timestamp>;
  email: string;
  id: Generated<string>;
  role: Generated<UserRole>;
  updated_at: Generated<Timestamp>;
}

export interface DB {
  campaigns: Campaigns;
  coupons: Coupons;
  redemptions: Redemptions;
  users: Users;
}